+++ testutil/integration/evmos/grpc/evm.go
@@ -4,8 +4,10 @@ package grpc
 
 import (
 	"context"
+	"errors"
 
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/evmos/evmos/v20/x/evm/core/vm"
 	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
 )
 
@@ -20,10 +22,23 @@ func (gqh *IntegrationHandler) GetEvmAccount(address common.Address) (*evmtypes.
 // EstimateGas returns the estimated gas for the given call args.
 func (gqh *IntegrationHandler) EstimateGas(args []byte, gasCap uint64) (*evmtypes.EstimateGasResponse, error) {
 	evmClient := gqh.network.GetEvmClient()
-	return evmClient.EstimateGas(context.Background(), &evmtypes.EthCallRequest{
+	res, err := evmClient.EstimateGas(context.Background(), &evmtypes.EthCallRequest{
 		Args:   args,
 		GasCap: gasCap,
 	})
+	if err != nil {
+		return nil, err
+	}
+
+	// handle case where there's a revert related error
+	if res.Failed() {
+		if (res.VmError != vm.ErrExecutionReverted.Error()) || len(res.Ret) == 0 {
+			return nil, errors.New(res.VmError)
+		}
+		return nil, evmtypes.NewExecErrorWithReason(res.Ret)
+	}
+
+	return res, err
 }
 
 // GetEvmParams returns the EVM module params.